# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SlipStream",
    "description": "Real-time social driving network for car enthusiasts.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/friends": {
      "get": {
        "tags": [
          "friends"
        ],
        "summary": "List Friends",
        "description": "List all accepted friends.",
        "operationId": "list_friends",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FriendProfile"
                  },
                  "type": "array",
                  "title": "Response Friends-List Friends"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/friends/requests": {
      "get": {
        "tags": [
          "friends"
        ],
        "summary": "List Friend Requests",
        "description": "List pending incoming friend requests.",
        "operationId": "list_friend_requests",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FriendRequestOut"
                  },
                  "type": "array",
                  "title": "Response Friends-List Friend Requests"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/friends/request": {
      "post": {
        "tags": [
          "friends"
        ],
        "summary": "Send Friend Request",
        "description": "Send a friend request to another user.",
        "operationId": "send_friend_request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FriendRequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/friends/accept": {
      "post": {
        "tags": [
          "friends"
        ],
        "summary": "Accept Friend Request",
        "description": "Accept an incoming friend request.",
        "operationId": "accept_friend_request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FriendRequestAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/friends/decline": {
      "post": {
        "tags": [
          "friends"
        ],
        "summary": "Decline Friend Request",
        "description": "Decline an incoming friend request.",
        "operationId": "decline_friend_request",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FriendRequestAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/friends/{user_id}": {
      "delete": {
        "tags": [
          "friends"
        ],
        "summary": "Remove Friend",
        "description": "Remove a friend (unfriend).",
        "operationId": "remove_friend",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/convoys/{convoy_id}/end": {
      "post": {
        "tags": [
          "convoys"
        ],
        "summary": "End Convoy",
        "description": "End the convoy (leader only). Dissolves it for everyone.",
        "operationId": "end_convoy",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "convoy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Convoy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/convoys/{convoy_id}/route": {
      "post": {
        "tags": [
          "convoys"
        ],
        "summary": "Set Route",
        "description": "Set a group route for the convoy (any member can set).",
        "operationId": "set_route",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "convoy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Convoy Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetRouteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "convoys"
        ],
        "summary": "Get Active Route",
        "description": "Get the current active route for the convoy.",
        "operationId": "get_active_route",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "convoy_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Convoy Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/RouteResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Convoys-Get Active Route"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FriendCarSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "year": {
            "type": "integer",
            "title": "Year"
          },
          "make": {
            "type": "string",
            "title": "Make"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "trim": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trim"
          },
          "color": {
            "type": "string",
            "title": "Color"
          }
        },
        "type": "object",
        "required": [
          "id",
          "year",
          "make",
          "model",
          "color"
        ],
        "title": "FriendCarSummary"
      },
      "FriendProfile": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "active_car": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FriendCarSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "username"
        ],
        "title": "FriendProfile"
      },
      "FriendRequestAction": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "request_id"
        ],
        "title": "FriendRequestAction"
      },
      "FriendRequestCreate": {
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "FriendRequestCreate"
      },
      "FriendRequestOut": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "from_user": {
            "$ref": "#/components/schemas/FriendProfile"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "from_user",
          "created_at"
        ],
        "title": "FriendRequestOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageResponse"
      },
...
      "RouteResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "destination_name": {
            "type": "string",
            "title": "Destination Name"
          },
          "destination_lat": {
            "type": "number",
            "title": "Destination Lat"
          },
          "destination_lng": {
            "type": "number",
            "title": "Destination Lng"
          },
          "set_by_username": {
            "type": "string",
            "title": "Set By Username"
          },
          "waypoints": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Waypoint"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Waypoints"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "destination_name",
          "destination_lat",
          "destination_lng",
          "set_by_username",
          "is_active",
          "created_at"
        ],
        "title": "RouteResponse"
      },
...
      "SetRouteRequest": {
        "properties": {
          "destination_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Destination Name"
          },
          "destination_lat": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Destination Lat"
          },
          "destination_lng": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Destination Lng"
          },
          "waypoints": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Waypoint"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Waypoints"
          }
        },
        "type": "object",
        "required": [
          "destination_name",
          "destination_lat",
          "destination_lng"
        ],
        "title": "SetRouteRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "Waypoint": {
        "properties": {
          "lat": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat"
          },
          "lng": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lng"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order"
          }
        },
        "type": "object",
        "required": [
          "lat",
          "lng"
        ],
        "title": "Waypoint",
        "description": "A single waypoint along a route."
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}